Fixed bug in setting number of scripted dof

+++ BASim/src/Physics/ElasticRods/RodTimeStepper.hh
@@ -425,7 +425,7 @@ public:
         = m_rod.getBoundaryCondition()->scriptedEdges();
       //= m_boundaryCondition->scriptedEdges();
 
-    int nb = 3 * (int)( verts.size() + edges.size() ); // # of scripted dofs
+    int nb = 3 * (int)verts.size() + (int)edges.size(); // # of scripted dofs 
     indices.resize(nb);
     desired.resize(nb);
